requirements.txt
torch
torchvision
diffusers
transformers
accelerate
pyyaml
Pillow
fonttools
uharfbuzz
scikit-image
tqdm
